requirements.txt
# GUARD - Generative Understanding for Anomaly Response & Detection
# Python Dependencies

# Deep Learning
tensorflow>=2.15.0
keras>=3.0.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# LLM Integration (for Root Cause Analysis)
openai>=1.0.0  # Groq API (OpenAI-compatible)

# Environment Configuration
python-dotenv>=1.0.0

# Email (built-in, no install needed)
# - smtplib (standard library)
# - email (standard library)

# PDF Generation
reportlab>=4.0.0

# Dashboard (Streamlit)
streamlit>=1.28.0
plotly>=5.17.0
